TODO:
- check Mtest output to approve tests or fix code
- update geotiff impl. to work with the new create array mechanism

+++ common/utils/msabaoth.c
@@ -51,6 +51,8 @@
 #define unlink _unlink
 #endif
 
+#define PATHLENGTH 4096
+
 /** the directory where the databases are (aka dbfarm) */
 char *_sabaoth_internal_dbfarm = NULL;
 /** the database which is "active" */
@@ -106,8 +108,8 @@ getDBPath(char **ret, size_t size, const char *extra)
  * dbname may be NULL to indicate that there is no active database.  The
  * arguments are copied for internal use.
  */
-void
-msab_init(char *dbfarm, char *dbname)
+static void
+msab_init(const char *dbfarm, const char *dbname)
 {
 	size_t len;
 
@@ -143,6 +145,23 @@ msab_init(char *dbfarm, char *dbname)
 		_sabaoth_internal_dbname = strdup(dbname);
 	}
 }
+void
+msab_dbpathinit(const char *dbpath)
+{
+	char dbfarm[PATHLENGTH];
+	const char *p;
+
+	p = strrchr(dbpath, DIR_SEP);
+	assert(p != NULL);
+	strncpy(dbfarm, dbpath, p - dbpath);
+	dbfarm[p - dbpath] = 0;
+	msab_init(dbfarm, p + 1);
+}
+void
+msab_dbfarminit(const char *dbfarm)
+{
+	msab_init(dbfarm, NULL);
+}
 
 /**
  * Returns the dbfarm as received during msab_init.  Returns an
@@ -171,8 +190,6 @@ msab_getDBname(char **ret)
 	*ret = strdup(_sabaoth_internal_dbname);
 	return(NULL);
 }
-
-#define PATHLENGTH 4096
 
 #define SCENARIOFILE ".scen"
 
@@ -230,7 +247,7 @@ char *
 msab_retreatScenario(const char *lang)
 {
 	FILE *f;
-	char buf[256];	/* should be enough for now */
+	char buf[256];	/* should be enough to hold the entire file */
 	size_t len;
 	char pathbuf[PATHLENGTH];
 	char *path = pathbuf;
@@ -242,30 +259,25 @@ msab_retreatScenario(const char *lang)
 	if ((f = fopen(path, "a+")) != NULL) {
 		if ((len = fread(buf, 1, 255, f)) > 0) {
 			char *p;
-			FILE *tmpf = tmpfile();
-			int written = 0;
+			char written = 0;
 
 			buf[len] = '\0';
 			tmp = buf;
 			/* find newlines and evaluate string */
 			while ((p = strchr(tmp, '\n')) != NULL) {
 				*p = '\0';
-				if (strcmp(tmp, lang) != 0) {
-					fprintf(tmpf, "%s\n", buf);
+				if (strcmp(tmp, lang) == 0) {
+					memmove(tmp, p + 1, strlen(p + 1) + 1);
 					written = 1;
+				} else {
+					*p = '\n';
+					tmp = p;
 				}
-				tmp = p;
 			}
 			if (written != 0) {
-				/* no idea how to "move" a file by it's fd (sounds
-				 * impossible anyway) and tmpnam is so much "DO NOT USE"
-				 * that I decided to just copy over the file again... */
 				rewind(f);
-				fflush(tmpf);
-				rewind(tmpf);
-				len = fread(tmp, 1, 256, tmpf);
-				if (fwrite(tmp, 1, len, f) < len) {
-					(void)fclose(tmpf);
+				len = strlen(buf) + 1;
+				if (fwrite(buf, 1, len, f) < len) {
 					(void)fclose(f);
 					snprintf(buf, sizeof(buf), "failed to write: %s (%s)",
 							strerror(errno), path);
@@ -273,7 +285,6 @@ msab_retreatScenario(const char *lang)
 				}
 				fflush(f);
 				fclose(f);
-				fclose(tmpf); /* this should remove it automagically */
 				return(NULL);
 			} else {
 				(void)fclose(f);
@@ -282,7 +293,7 @@ msab_retreatScenario(const char *lang)
 			}
 		} else if (len == 0) {
 			(void)fclose(f);
-			unlink(path);
+			unlink(path);  /* empty file? try to remove */
 			return(NULL);
 		} else { /* some error */
 			(void)fclose(f);
@@ -337,6 +348,7 @@ msab_marchConnection(const char *host, const int port)
 	}
 }
 
+#define STARTEDFILE ".started"
 /**
  * Removes all known publications of available services.  The function
  * name is a nostalgic phrase from "Defender of the Crown" from the
@@ -357,6 +369,10 @@ msab_wildRetreat(void)
 		return(tmp);
 	unlink(path);
 
+	if ((tmp = getDBPath(&path, PATHLENGTH, STARTEDFILE)) != NULL)
+		return(tmp);
+	unlink(path);
+
 	if ((tmp = getDBPath(&path, PATHLENGTH, _sabaoth_internal_uuid)) != NULL)
 		return(tmp);
 	unlink(path);
@@ -365,7 +381,6 @@ msab_wildRetreat(void)
 }
 
 #define UPLOGFILE ".uplog"
-#define STARTINGFILE ".starting"
 /**
  * Writes a start attempt to the sabaoth start/stop log.  Examination of
  * the log at a later stage reveals crashes of the server.  In addition
@@ -409,11 +424,12 @@ msab_registerStarting(void)
 		return(NULL);
 	}
 	fclose(fopen(path, "w"));
-	/* flag this database as starting up, with the same boundary
-	 * conditions as above */
-	if ((tmp = getDBPath(&path, PATHLENGTH, STARTINGFILE)) != NULL)
+	
+	/* remove any stray file that would suggest we've finished starting up */
+	if ((tmp = getDBPath(&path, PATHLENGTH, STARTEDFILE)) != NULL)
 		return(tmp);
-	fclose(fopen(path, "w"));
+	unlink(path);
+
 
 	return(NULL);
 }
@@ -430,10 +446,10 @@ msab_registerStarted(void)
 	char *path = pathbuf;
 	char *tmp;
 
-	/* remove starting flag */
-	if ((tmp = getDBPath(&path, PATHLENGTH, STARTINGFILE)) != NULL)
+	/* flag this database as started up */
+	if ((tmp = getDBPath(&path, PATHLENGTH, STARTEDFILE)) != NULL)
 		return(tmp);
-	unlink(path);
+	fclose(fopen(path, "w"));
 
 	return(NULL);
 }
@@ -549,6 +565,7 @@ msab_getStatus(sabdb** ret, char *dbname)
 			sdb = sdb->next = malloc(sizeof(sabdb));
 		}
 		sdb->uplog = NULL;
+		sdb->uri = NULL;
 		sdb->next = NULL;
 
 		/* store the database name */
@@ -625,11 +642,11 @@ msab_getStatus(sabdb** ret, char *dbname)
 				} else if (data[0] == '\t') {
 					/* see if the database has finished starting */
 					snprintf(buf, sizeof(buf), "%s/%s/%s",
-							path, e->d_name, STARTINGFILE);
+							path, e->d_name, STARTEDFILE);
 					if (stat(buf, &statbuf) == -1) {
-						sdb->state = SABdbRunning;
-					} else {
 						sdb->state = SABdbStarting;
+					} else {
+						sdb->state = SABdbRunning;
 					}
 				} else { /* should be \n */
 					sdb->state = SABdbInactive;
@@ -648,11 +665,11 @@ msab_getStatus(sabdb** ret, char *dbname)
 			/* lock denied, so mserver is running, see if the database
 			 * has finished starting */
 			snprintf(buf, sizeof(buf), "%s/%s/%s",
-					path, e->d_name, STARTINGFILE);
+					path, e->d_name, STARTEDFILE);
 			if (stat(buf, &statbuf) == -1) {
-				sdb->state = SABdbRunning;
-			} else {
 				sdb->state = SABdbStarting;
+			} else {
+				sdb->state = SABdbRunning;
 			}
 		} else {
 			/* locking succeed, check for a crash in the uplog */
@@ -700,6 +717,8 @@ msab_freeStatus(sabdb** ret)
 	while (p != NULL) {
 		if (p->path != NULL)
 			free(p->path);
+		if (p->uri != NULL)
+			free(p->uri);
 		if (p->uplog != NULL)
 			free(p->uplog);
 		r = p->scens;
@@ -755,6 +774,7 @@ msab_getUplogInfo(sabuplog *ret, const sabdb *db)
 	memset(ret, 0, sizeof(sabuplog));
 	ret->minuptime = -1;
 	ret->lastcrash = -1;
+	ret->laststop = -1;
 
 	snprintf(log, sizeof(log), "%s/%s", db->path, UPLOGFILE);
 	if ((f = fopen(log, "r")) != NULL) {
@@ -781,7 +801,7 @@ msab_getUplogInfo(sabuplog *ret, const sabdb *db)
 					/* successful stop */
 					ret->stopcntr++;
 					*p = '\0';
-					stop = (time_t)atol(data);
+					ret->laststop = stop = (time_t)atol(data);
 					p = data;
 					i = (int) (stop - start);
 					if (i > ret->maxuptime)
@@ -837,7 +857,7 @@ msab_getUplogInfo(sabuplog *ret, const sabdb *db)
 }
 
 /* used in the serialisation to be able to change it in the future */
-#define SABDBVER "1"
+#define SABDBVER "2"
 
 /**
  * Produces a string representation suitable for storage/sending.
@@ -847,7 +867,6 @@ msab_serialise(char **ret, const sabdb *db)
 {
 	char buf[8096];
 	char scens[64];
-	char conns[1024];
 	sablist *l;
 	sabuplog dbu;
 	char *p;
@@ -869,35 +888,21 @@ msab_serialise(char **ret, const sabdb *db)
 	if (p != scens)
 		p[-1] = '\0';
 
-	conns[0] = '\0';
-	p = conns;
-	avail = sizeof(conns) - 1;
-	for (l = db->conns; l != NULL; l = l->next) {
-		len = strlen(l->val);
-		if (len > avail)
-			break;
-		memcpy(p, l->val, len);
-		p += len + 1;
-		avail -= len + 1;
-		memcpy(p - 1, "'", 2);
-	}
-	if (p != conns)
-		p[-1] = '\0';
-
 	if ((p = msab_getUplogInfo(&dbu, db)) != NULL)
 		return(p);
 
 	/* sabdb + sabuplog structs in one */
 	snprintf(buf, sizeof(buf), "sabdb:" SABDBVER ":"
-			"%s,%d,%d,%s,%s" ","
+			"%s,%s,%d,%d,%s,"
 			"%d,%d,%d,"
 			"" LLFMT "," LLFMT "," LLFMT ","
-			"" LLFMT "," LLFMT ","
+			"" LLFMT "," LLFMT "," LLFMT ","
 			"%d,%f,%f",
-			db->path, db->locked, (int)(db->state), scens, conns,
+			db->dbname, db->uri ? db->uri : "", db->locked,
+			(int)(db->state), scens,
 			dbu.startcntr, dbu.stopcntr, dbu.crashcntr,
 			(lng)dbu.avguptime, (lng)dbu.maxuptime, (lng)dbu.minuptime,
-			(lng)dbu.lastcrash, (lng)dbu.laststart,
+			(lng)dbu.lastcrash, (lng)dbu.laststart, (lng)dbu.laststop,
 			dbu.crashavg1, dbu.crashavg10, dbu.crashavg30);
 
 	*ret = strdup(buf);
@@ -910,17 +915,18 @@ msab_serialise(char **ret, const sabdb *db)
 char *
 msab_deserialise(sabdb **ret, char *sdb)
 {
-	char *path;
+	char *dbname;
+	char *uri;
 	int locked;
 	int state;
 	char *scens = "";
-	char *conns = "";
 	sabdb *s;
 	sabuplog *u;
 	sablist *l;
 	char *p;
 	char *lasts;
 	char buf[PATHLENGTH];
+	char protover = 0;
 
 	lasts = sdb;
 	if ((p = strchr(lasts, ':')) == NULL) {
@@ -941,19 +947,46 @@ msab_deserialise(sabdb **ret, char *sdb)
 		return(strdup(buf));
 	}
 	*p++ = '\0';
-	if (strcmp(lasts, SABDBVER) != 0) {
+	if (strcmp(lasts, "1") == 0) {
+		/* Protocol 1 was used uptil Oct2012.  Since Jul2012 a new state
+		 * SABdbStarting was introduced, but not exposed to the client
+		 * in serialise.  In Feb2013, the path component was removed
+		 * and replaced by an URI field.  This meant dbname could no
+		 * longer be deduced from path, and hence sent separately.
+		 * Since the conns property became useless in the light of the
+		 * added uri, it was dropped.  On top of this, a laststop
+		 * property was added to the uplog struct.
+		 * These four changes were effectuated in protocol 2.  When
+		 * reading protocol 1, we use the path field to set dbname, but
+		 * ignore the path information (and set uri to "<unknown>".  The
+		 * SABdbStarting state never occurs. */
+	} else if (strcmp(lasts, SABDBVER) != 0) {
 		snprintf(buf, sizeof(buf), 
 				"string has unsupported version: %s", lasts);
 		return(strdup(buf));
 	}
+	protover = lasts[0];
 	lasts = p;
 	if ((p = strchr(p, ',')) == NULL) {
 		snprintf(buf, sizeof(buf), 
-				"string does not contain path: %s", lasts);
+				"string does not contain %s: %s",
+				protover == '1' ? "path" : "dbname", lasts);
 		return(strdup(buf));
 	}
 	*p++ = '\0';
-	path = lasts;
+	dbname = lasts;
+	if (protover == '1') {
+		uri = "<unknown>";
+	} else {
+		lasts = p;
+		if ((p = strchr(p, ',')) == NULL) {
+			snprintf(buf, sizeof(buf), 
+					"string does not contain uri: %s", lasts);
+			return(strdup(buf));
+		}
+		*p++ = '\0';
+		uri = lasts;
+	}
 	lasts = p;
 	if ((p = strchr(p, ',')) == NULL) {
 		snprintf(buf, sizeof(buf), 
@@ -979,19 +1012,21 @@ msab_deserialise(sabdb **ret, char *sdb)
 	*p++ = '\0';
 	scens = lasts;
 	lasts = p;
-	if ((p = strchr(p, ',')) == NULL) {
-		snprintf(buf, sizeof(buf), 
-				"string does not contain connections: %s", lasts);
-		return(strdup(buf));
+	if (protover == '1') {
+		if ((p = strchr(p, ',')) == NULL) {
+			snprintf(buf, sizeof(buf), 
+					"string does not contain connections: %s", lasts);
+			return(strdup(buf));
+		}
+		*p++ = '\0';
+		lasts = p;
 	}
-	*p++ = '\0';
-	conns = lasts;
-	lasts = p;
 
 	/* start parsing sabuplog struct */
 	u = malloc(sizeof(sabuplog));
 
 	if ((p = strchr(p, ',')) == NULL) {
+		free(u);
 		snprintf(buf, sizeof(buf), 
 				"string does not contain startcounter: %s", lasts);
 		return(strdup(buf));
@@ -1000,6 +1035,7 @@ msab_deserialise(sabdb **ret, char *sdb)
 	u->startcntr = atoi(lasts);
 	lasts = p;
 	if ((p = strchr(p, ',')) == NULL) {
+		free(u);
 		snprintf(buf, sizeof(buf), 
 				"string does not contain stopcounter: %s", lasts);
 		return(strdup(buf));
@@ -1008,6 +1044,7 @@ msab_deserialise(sabdb **ret, char *sdb)
 	u->stopcntr = atoi(lasts);
 	lasts = p;
 	if ((p = strchr(p, ',')) == NULL) {
+		free(u);
 		snprintf(buf, sizeof(buf), 
 				"string does not contain crashcounter: %s", lasts);
 		return(strdup(buf));
@@ -1016,6 +1053,7 @@ msab_deserialise(sabdb **ret, char *sdb)
 	u->crashcntr = atoi(lasts);
 	lasts = p;
 	if ((p = strchr(p, ',')) == NULL) {
+		free(u);
 		snprintf(buf, sizeof(buf), 
 				"string does not contain avguptime: %s", lasts);
 		return(strdup(buf));
@@ -1024,6 +1062,7 @@ msab_deserialise(sabdb **ret, char *sdb)
 	u->avguptime = (time_t)strtoll(lasts, (char **)NULL, 10);
 	lasts = p;
 	if ((p = strchr(p, ',')) == NULL) {
+		free(u);
 		snprintf(buf, sizeof(buf), 
 				"string does not contain maxuptime: %s", lasts);
 		return(strdup(buf));
@@ -1032,6 +1071,7 @@ msab_deserialise(sabdb **ret, char *sdb)
 	u->maxuptime = (time_t)strtoll(lasts, (char **)NULL, 10);
 	lasts = p;
 	if ((p = strchr(p, ',')) == NULL) {
+		free(u);
 		snprintf(buf, sizeof(buf), 
 				"string does not contain minuptime: %s", lasts);
 		return(strdup(buf));
@@ -1040,6 +1080,7 @@ msab_deserialise(sabdb **ret, char *sdb)
 	u->minuptime = (time_t)strtoll(lasts, (char **)NULL, 10);
 	lasts = p;
 	if ((p = strchr(p, ',')) == NULL) {
+		free(u);
 		snprintf(buf, sizeof(buf), 
 				"string does not contain lastcrash: %s", lasts);
 		return(strdup(buf));
@@ -1048,6 +1089,7 @@ msab_deserialise(sabdb **ret, char *sdb)
 	u->lastcrash = (time_t)strtoll(lasts, (char **)NULL, 10);
 	lasts = p;
 	if ((p = strchr(p, ',')) == NULL) {
+		free(u);
 		snprintf(buf, sizeof(buf), 
 				"string does not contain laststart: %s", lasts);
 		return(strdup(buf));
@@ -1055,7 +1097,21 @@ msab_deserialise(sabdb **ret, char *sdb)
 	*p++ = '\0';
 	u->laststart = (time_t)strtoll(lasts, (char **)NULL, 10);
 	lasts = p;
+	if (protover != '1') {
+		if ((p = strchr(p, ',')) == NULL) {
+			free(u);
+			snprintf(buf, sizeof(buf), 
+					"string does not contain laststop: %s", lasts);
+			return(strdup(buf));
+		}
+		*p++ = '\0';
+		u->laststop = (time_t)strtoll(lasts, (char **)NULL, 10);
+		lasts = p;
+	} else {
+		u->laststop = -1;
+	}
 	if ((p = strchr(p, ',')) == NULL) {
+		free(u);
 		snprintf(buf, sizeof(buf), 
 				"string does not contain crashavg1: %s", lasts);
 		return(strdup(buf));
@@ -1064,6 +1120,7 @@ msab_deserialise(sabdb **ret, char *sdb)
 	u->crashavg1 = atoi(lasts);
 	lasts = p;
 	if ((p = strchr(p, ',')) == NULL) {
+		free(u);
 		snprintf(buf, sizeof(buf), 
 				"string does not contain crashavg10: %s", lasts);
 		return(strdup(buf));
@@ -1072,6 +1129,7 @@ msab_deserialise(sabdb **ret, char *sdb)
 	u->crashavg10 = atof(lasts);
 	lasts = p;
 	if ((p = strchr(p, ',')) != NULL) {
+		free(u);
 		snprintf(buf, sizeof(buf), 
 				"string does contain additional garbage after crashavg30: %s",
 				lasts);
@@ -1081,16 +1139,20 @@ msab_deserialise(sabdb **ret, char *sdb)
 
 	/* fill/create sabdb struct */
 
-	if (strrchr(path, '/') == NULL) {
-		snprintf(buf, sizeof(buf), "invalid path: %s", path);
-		return(strdup(buf));
+	if (protover == '1') {
+		if ((dbname = strrchr(dbname, '/')) == NULL) {
+			free(u);
+			snprintf(buf, sizeof(buf), "invalid path: %s", dbname);
+			return(strdup(buf));
+		}
+		dbname++;
 	}
 
 	s = malloc(sizeof(sabdb));
 
-	s->path = strdup(path);
 	/* msab_freeStatus() actually relies on this trick */
-	s->dbname = strrchr(s->path, '/') + 1;
+	s->path = s->dbname = strdup(dbname);
+	s->uri = strdup(uri);
 	s->locked = locked;
 	s->state = (SABdbState)state;
 	if (strlen(scens) == 0) {
@@ -1111,24 +1173,7 @@ msab_deserialise(sabdb **ret, char *sdb)
 			}
 		}
 	}
-	if (strlen(conns) == 0) {
-		s->conns = NULL;
-	} else {
-		l = s->conns = malloc(sizeof(sablist));
-		p = strtok_r(conns, "'", &lasts);
-		if (p == NULL) {
-			l->val = strdup(conns);
-			l->next = NULL;
-		} else {
-			l->val = strdup(p);
-			l->next = NULL;
-			while ((p = strtok_r(NULL, "'", &lasts)) != NULL) {
-				l = l->next = malloc(sizeof(sablist));
-				l->val = strdup(p);
-				l->next = NULL;
-			}
-		}
-	}
+	s->conns = NULL;
 	s->uplog = u;
 	s->next = NULL;
